The interface is taking shape, I now kindof understand how jQuery works (yeah, kindof.. missing some basis in JavaScript etc..).
Fixed the model for User (forgot to add 'credit', can you believe it?). Anyone know if you can add a field to your database (using manage.py) without trashing your database?

Next in line is the actual submission of the info when a button is pushed, there is a small stub in views that starts implementing that, also implemented it in urls.py. So next time: get to know and love AJAX.

(oh, and exams ;-) )